Моковые функции вынесены в отдельный файл. Реализовано открытие/закрытие формы редактирования изображения.
