+++ js/main.js
@@ -1,5 +1,6 @@
 import {generatePhotosData} from './mocks/photos-data.js';
 import {renderMiniatures} from './miniatures.js';
+import './form.js';
 
 const COUNT_OF_PHOTOS = 25;
